+++ sdk/model/src/components/PlAnnotations/filter.ts
@@ -1,3 +1,4 @@
+// @DEPRECATED - use sdk/model/src/filters + sdk/model/src/annotations
 import type { SUniversalPColumnId } from '@milaboratories/pl-model-common';
 //
 // Sequence filter