+++ sdk/ui-vue/src/components/PlBtnExportArchive/index.ts
@@ -1,0 +1,2 @@
+export { default as PlBtnExportArchive } from './PlBtnExportArchive.vue';
+export type { FileExportEntry } from './types';