+++ etc/blocks/monetization-test/workflow/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @milaboratories/milaboratories.monetization-test.workflow
 
+## 1.1.31
+
+### Patch Changes
+
+- Updated dependencies [033f2fa]
+  - @platforma-sdk/workflow-tengo@5.5.1
+
+## 1.1.30
+
+### Patch Changes
+
+- Updated dependencies [3ef2381]
+  - @platforma-sdk/workflow-tengo@5.5.0
+
 ## 1.1.29
 
 ### Patch Changes