+++ sdk/workflow-tengo/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @platforma-sdk/workflow-tengo
 
+## 5.5.1
+
+### Patch Changes
+
+- 033f2fa: Add ParquetPartitioned format support to all workflow-tengo functions
+
+## 5.5.0
+
+### Minor Changes
+
+- 3ef2381: Generelazation filters and annotations
+
+### Patch Changes
+
+- Updated dependencies [3ef2381]
+  - @platforma-open/milaboratories.software-ptabler@1.13.0
+  - @platforma-open/milaboratories.software-ptexter@1.2.0
+
 ## 5.4.5
 
 ### Patch Changes